Move cgetArticles call to ArticleGrid component
