+++ src/ArticleGrid/ArticleGrid.js
@@ -1,7 +1,37 @@
+import { useEffect, useState } from 'react';
+import { getHome, getArticles } from '../calls';
 import { ArticleCard } from '../ArticleCard/ArticleCard';
 import './ArticleGrid.css';
 
-export const ArticleGrid = ({ articles }) => {
+export const ArticleGrid = ({ section }) => {
+  const [sectionArticles, setSectionArticles] = useState([]);
+
+  useEffect(() => {
+    cleanGETbySection();
+  }, [section])
+
+  const cleanGETbySection = () => {
+    getArticles(section)
+      .then(res => {
+        // console.log(res);
+        if (!res.ok) throw new Error();
+        return res.json();
+      })
+      .then(goodRes => {
+        // console.log(goodRes.results);
+        if (section) {
+          setSectionArticles(goodRes.results);
+        }
+
+        if (!section) {
+          setSectionArticles('home');
+        }
+      })
+      .catch((badRes) => {
+        // console.log(badRes);
+        return badRes;
+      })
+  }
 
   const chooseMediaSize = (options, desiredFormat) => {
     if (!options) return 'nyt.png';
@@ -10,10 +40,10 @@ export const ArticleGrid = ({ articles }) => {
   }
   
   const renderCards = () => {
-    return articles.map(art => {
+    return sectionArticles.map(art => {
       return (
         <ArticleCard 
-          media={ chooseMediaSize(art.multimedia, 'superJumbo') }
+          media={ chooseMediaSize(art.multimedia, 'Normal') }
           title={ art.title }
         />
       )
